requirements.txt
streamlit>=1.32.0
langgraph==0.2.28
langchain-groq==0.2.0
langchain-core==0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.19.0
pandas>=2.2.0
numpy>=1.26.0
